Move control js files into separate folder.
