+++ controls/BugReportControl.js
