+++ controls/InfoControl.js
