+++ controls/ZoomToPosControl.js
